create new token handle function
