+++ src/api/middlewares/setMainMiddleware.go
@@ -8,18 +8,10 @@ import (
 )
 
 func SetMainMiddlewares(e *echo.Echo) {
-	e.Use(serverHeader)
+	//cors対策
 	e.Use(middleware.CORSWithConfig(middleware.CORSConfig{
-        AllowOrigins: []string{"http://localhost:3000", "https://next-redux-app-tau.vercel.app/"},
-        AllowMethods: []string{http.MethodGet, http.MethodPut, http.MethodPost, http.MethodDelete},
-    }))
-}
-
-func serverHeader(next echo.HandlerFunc) echo.HandlerFunc {
-	return func(c echo.Context) error {
-		c.Response().Header().Set(echo.HeaderServer, "BlueBot/1.0")
-		c.Response().Header().Set("notReallyHeader", "thisHaveNoMeaning")
-
-		return next(c)
-	}
+		AllowOrigins:     []string{"http://localhost:3000", "https://next-redux-app-tau.vercel.app/"},
+		AllowMethods:     []string{http.MethodGet, http.MethodPut, http.MethodPost, http.MethodDelete, echo.HeaderOrigin, echo.HeaderContentType, echo.HeaderAccept},
+		AllowCredentials: true,
+	}))
 }
