--- src/firebaase-admin/snippets/auth.go
@@ -1,23 +1,0 @@
-package snippets
-
-import (
-	"context"
-	"log"
-	"firebase.google.com/go/v4/auth"
-	firebase "firebase.google.com/go/v4"
-)
-
-func VerifyIDToken(ctx context.Context, app *firebase.App, idToken string) *auth.Token {
-	client, err := app.Auth(ctx)
-	if err != nil {
-		log.Fatalf("error getting Auth client: %v\n", err)
-	}
-
-	token, err := client.VerifyIDToken(ctx, idToken)
-	if err != nil {
-		log.Fatalf("error verifying ID token: %v\n", err)
-	}
-
-	log.Printf("Verified ID token: %v\n", token)
-	return token
-}
