+++ src/main/main.go
@@ -3,8 +3,8 @@ package main
 import (
 	"fmt"
 	// "log"
+	"myapp/src/api/firebase/snippets"
 	"myapp/src/db"
-	"myapp/src/firebaase-admin/snippets"
 	"myapp/src/router"
 
 	_ "github.com/jinzhu/gorm/dialects/mysql"
